+++ sshd-common/src/test/java/org/apache/sshd/common/cipher/AES192CTRTest.java
@@ -35,7 +35,7 @@ public class AES192CTRTest extends BaseCipherTest {
     @Test
     public void testEncryptDecrypt() throws Exception {
         // for AES 256 bits we need the JCE unlimited strength policy
-        ensureKeySizeSupported(16, 24, "AES", "AES/CTR/NoPadding");
+        ensureFullCipherInformationSupported(BuiltinCiphers.aes192ctr);
         testEncryptDecrypt(BuiltinCiphers.aes192ctr);
     }
 }