+++ sshd-core/src/main/java/org/apache/sshd/common/io/PacketWriter.java
@@ -46,11 +46,36 @@ public interface PacketWriter extends Channel {
      * @return The required padding length
      */
     static int calculatePadLength(int len, int blockSize, boolean etmMode) {
+        /*
+         * Note: according to RFC-4253 section 6:
+         *
+         *    The minimum size of a packet is 16 (or the cipher block size,
+         *     whichever is larger) bytes (plus 'mac').
+         *
+         * Since all out ciphers, MAC(s), etc. have a block size > 8 then
+         * the minimum size of the packet will be at least 16 due to the
+         * padding at the very least - so even packets that contain an opcode
+         * with no arguments will be above this value. This avoids an un-necessary
+         * call to Math.max(len, 16) for each and every packet
+         */
+
         len++;  // the pad length
         if (!etmMode) {
             len += Integer.BYTES;
         }
 
+        /*
+         * Note: according to RFC-4253 section 6:
+         *
+         *      Note that the length of the concatenation of 'packet_length',
+         *      'padding_length', 'payload', and 'random padding' MUST be a multiple
+         *      of the cipher block size or 8, whichever is larger.
+         *
+         * However, we currently do not have ciphers with a block size of less than
+         * 8 so we do not take this into account in order to accelerate the calculation
+         * and avoiding an un-necessary call to Math.max(blockSize, 8) for each and every
+         * packet.
+         */
         int pad = (-len) & (blockSize - 1);
         if (pad < blockSize) {
             pad += blockSize;